refactor parallel.
