+++ data/gan/datasets/default_cyclegan_dataset.py
@@ -8,7 +8,7 @@ import os
 import random
 import torch.utils.data as data
 
-from lib.exts.tools.parallel.data_container import DataContainer
+from lib.parallel.data_container import DataContainer
 from lib.tools.helper.image_helper import ImageHelper
