+++ data/gan/datasets/default_facegan_dataset.py
@@ -9,7 +9,7 @@ import json
 import random
 import torch.utils.data as data
 
-from lib.exts.tools.parallel.data_container import DataContainer
+from lib.parallel.data_container import DataContainer
 from lib.tools.helper.image_helper import ImageHelper
 from lib.tools.util.logger import Logger as Log
