+++ data/pose/datasets/default_openpose_dataset.py
@@ -9,7 +9,7 @@ import torch
 import numpy as np
 import torch.utils.data as data
 
-from lib.exts.tools.parallel.data_container import DataContainer
+from lib.parallel.data_container import DataContainer
 from lib.tools.helper.json_helper import JsonHelper
 from lib.tools.helper.image_helper import ImageHelper
 from lib.tools.util.logger import Logger as Log
