+++ data/seg/datasets/cityscapes_dataset.py
@@ -7,7 +7,7 @@ import os
 import numpy as np
 from torch.utils import data
 
-from lib.exts.tools.parallel.data_container import DataContainer
+from lib.parallel.data_container import DataContainer
 from lib.tools.helper.image_helper import ImageHelper
 from lib.tools.helper.file_helper import FileHelper
 from lib.tools.util.logger import Logger as Log
