+++ lib/parallel/README.md
