+++ model/cls/model_manager.py
@@ -35,5 +35,5 @@ class ModelManager(object):
         if self.configer.get('network', 'gather'):
             return Loss(self.configer)
 
-        from lib.exts.tools.parallel.data_parallel import ParallelCriterion
+        from lib.parallel.data_parallel import ParallelCriterion
         return ParallelCriterion(Loss(self.configer))
