+++ model/seg/model_manager.py
@@ -40,5 +40,5 @@ class ModelManager(object):
         if self.configer.get('network', 'gather'):
             return Loss(self.configer)
 
-        from lib.exts.tools.parallel.data_parallel import ParallelCriterion
+        from lib.parallel.data_parallel import ParallelCriterion
         return ParallelCriterion(Loss(self.configer))
